+++ tests/coordination/test_leaderelection.py
@@ -10,7 +10,7 @@ from pytest_mock import MockerFixture
 
 import grelmicro.coordination._base as base_module
 import grelmicro.coordination.leaderelection as le_module
-from grelmicro.coordination.abc import LeaderElectionBackend
+from grelmicro.coordination._protocol import LeaderElectionBackend
 from grelmicro.coordination.errors import CoordinationSettingsValidationError
 from grelmicro.coordination.leaderelection import (
     LeaderElection,